+++ xc/lib/Xaw/Text.c
@@ -1267,7 +1267,7 @@ XtPointer callData;		/* #pixels */
 {
   TextWidget ctx = (TextWidget) closure;
   Widget tw = (Widget) ctx;
-  Position old_left, pixels = (Position)(int) callData;
+  Position old_left, pixels = (Position)(long) callData;
   XRectangle rect, t_rect;
   
   _XawTextPrepareToUpdate(ctx);
@@ -1431,7 +1431,7 @@ XtPointer closure;		/* TextWidget */
 XtPointer callData;		/* #pixels */
 {
   TextWidget ctx = (TextWidget)closure;
-  int height, lines = (int) callData;
+  int height, lines = (long) callData;
 
   height = ctx->core.height - VMargins(ctx);
   if (height < 1)
